+++ webapp-frontend/src/app/services/ui/instance-selector-dialog.service.spec.ts
@@ -2,14 +2,14 @@
  * ========================LICENSE_START=================================
  * O-RAN-SC
  * %%
- * Copyright (C) 2019 AT&T Intellectual Property
+ * Copyright (C) 2020 AT&T Intellectual Property
  * %%
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,28 +17,15 @@
  * limitations under the License.
  * ========================LICENSE_END===================================
  */
-import { Component, OnInit, Output, EventEmitter } from '@angular/core';
-import { UiService } from '../../services/ui/ui.service';
+import { TestBed } from '@angular/core/testing';
 
-@Component({
-  selector: 'rd-sidenav-list',
-  templateUrl: './sidenav-list.component.html',
-  styleUrls: ['./sidenav-list.component.scss']
-})
-export class SidenavListComponent implements OnInit {
-  darkMode: boolean;
-  @Output() sidenavClose: EventEmitter<any>  = new EventEmitter();
+import { InstanceSelectorDialogService } from './instance-selector-dialog.service';
 
-  constructor(public ui: UiService) { }
+describe('InstanceSelectorDialogService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
 
-  ngOnInit() {
-    this.ui.darkModeState.subscribe((isDark) => {
-      this.darkMode = isDark;
-    });
-  }
-
-  public onSidenavClose = () => {
-    this.sidenavClose.emit();
-  }
-
-}
+  it('should be created', () => {
+    const service: InstanceSelectorDialogService = TestBed.get(InstanceSelectorDialogService);
+    expect(service).toBeTruthy();
+  });
+});