+++ webapp-frontend/src/app/services/ui/instance-selector-dialog.service.ts
@@ -2,14 +2,14 @@
  * ========================LICENSE_START=================================
  * O-RAN-SC
  * %%
- * Copyright (C) 2019 AT&T Intellectual Property
+ * Copyright (C) 2020 AT&T Intellectual Property
  * %%
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,28 +17,36 @@
  * limitations under the License.
  * ========================LICENSE_END===================================
  */
-import { Component, OnInit, Output, EventEmitter } from '@angular/core';
-import { UiService } from '../../services/ui/ui.service';
+import { Injectable } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { InstanceSelectorDialogComponent } from '../../ui/instance-selector-dialog/instance-selector-dialog.component';
+import { UiService } from './ui.service';
 
-@Component({
-  selector: 'rd-sidenav-list',
-  templateUrl: './sidenav-list.component.html',
-  styleUrls: ['./sidenav-list.component.scss']
+
+@Injectable({
+  providedIn: 'root'
 })
-export class SidenavListComponent implements OnInit {
+export class InstanceSelectorDialogService {
+
   darkMode: boolean;
-  @Output() sidenavClose: EventEmitter<any>  = new EventEmitter();
+  panelClass: string;
 
-  constructor(public ui: UiService) { }
+  constructor(private dialog: MatDialog,
+    public ui: UiService) { }
 
-  ngOnInit() {
+  openInstanceSelectorDialog() {
     this.ui.darkModeState.subscribe((isDark) => {
       this.darkMode = isDark;
     });
+    if (this.darkMode) {
+      this.panelClass = 'dark-theme';
+    } else {
+      this.panelClass = '';
+    }
+    return this.dialog.open(InstanceSelectorDialogComponent, {
+      panelClass: this.panelClass,
+      width: '480px',
+      position: { top: '100px' },
+    });
   }
-
-  public onSidenavClose = () => {
-    this.sidenavClose.emit();
-  }
-
 }